+++ src/pages/PrivacyPolicy.js
@@ -1,0 +1,10 @@
+import ScreenContainer from "../components/screen";
+
+export default function PrivacyPolicy() {
+    return (
+      <ScreenContainer>
+        
+      </ScreenContainer>
+    );
+  }
+
