+++ src/pages/TermsAndConditions.js
@@ -1,0 +1,9 @@
+import ScreenContainer from "../components/screen";
+
+export default function TermsAndConditions() {
+  return (
+    <ScreenContainer>
+        
+    </ScreenContainer>
+  );
+}
